Backend/requirements.txt
flask-security-too
Flask-Cors
flask-restful
flask 
Flask-SQLAlchemy
setuptools
bcrypt==4.0.1
argon2_cffi